This was useful while [investigating basebackup slowness]( https://www.notion.so/neondatabase/Christian-Investigation-Slow-Basebackups-Early-2023-12-34ea5c7dcdc1485d9ac3731da4d2a6fc?pvs=4)

For posterity, a branch on top of this PR that contains additional debug spans used in that investigation: `problame/2023-12--slow-basebackup-investigation--cleaned-up2` (we might never commit it, the debug spans inhibit readability in the code).