@problame 2023-12-13
I'm not sure whether the use of `Box<dyn ...>`, without enabling the two layers introduced in this PR, has substantial performance impact at runtime.

Will test that with my WIP getpage@lsn benchmark #5771 

If it doesn't, I'm inclined to merge this.
@koivunej is that ok for you or can you not bear the boxes even if they don't have a measurable perf impact?