@koivunej 2023-12-20
Following the previous comment we discussed this and Christian was going to bench the effects.