@problame 2023-12-20
```
# setup
./target/release/neon_local init
./target/release/neon_local start
./target/release/neon_local tenant create --set-default

# workload
./target/release/pagebench get-page-latest-lsn --runtime 10s         
```

(`pagebench` from https://github.com/neondatabase/neon/pull/6174 )

Results on an idle i3en.3xlarge (`test_output` on EBS volume though, not the local NVME)

The tenant has an initdb'd timeline with no other contents.
The entire working set fits in RAM => no IO.

Without this PR

```
2023-12-20T16:49:36.665603Z  INFO RPS: 17360
2023-12-20T16:49:37.666599Z  INFO RPS: 17617
2023-12-20T16:49:38.667604Z  INFO RPS: 17318
2023-12-20T16:49:39.668606Z  INFO RPS: 17267
2023-12-20T16:49:40.669594Z  INFO RPS: 17517
2023-12-20T16:49:41.670623Z  INFO RPS: 17263
2023-12-20T16:49:42.671599Z  INFO RPS: 17382
2023-12-20T16:49:43.672599Z  INFO RPS: 17369
2023-12-20T16:49:44.673621Z  INFO RPS: 17317
{
  "total": {
    "request_count": 173891,
    "latency_mean": "56us",
    "latency_percentiles": {
      "p95": "67us",
      "p99": "67us",
      "p99.9": "67us",
      "p99.99": "67us"
    }
  }
}
```

With this PR:

```
2023-12-20T16:50:45.669375Z  INFO RPS: 15021
2023-12-20T16:50:46.670368Z  INFO RPS: 15092
2023-12-20T16:50:47.671358Z  INFO RPS: 14725
2023-12-20T16:50:48.672378Z  INFO RPS: 15096
2023-12-20T16:50:49.673377Z  INFO RPS: 14882
2023-12-20T16:50:50.674375Z  INFO RPS: 15020
2023-12-20T16:50:51.675371Z  INFO RPS: 15304
2023-12-20T16:50:52.676364Z  INFO RPS: 15063
2023-12-20T16:50:53.677372Z  INFO RPS: 15070
{
  "total": {
    "request_count": 150402,
    "latency_mean": "65us",
    "latency_percentiles": {
      "p95": "80us",
      "p99": "80us",
      "p99.9": "80us",
      "p99.99": "80us"
    }
  }
}
```


So, I guess there's runtime overhead and we can't do it like it is in this PR.