@problame 2025-01-07
Need to talk to the Cargo wizards to figure out whether&how we can avoid compile bloat.